list pytorch device correctly for pytest
